Modifications to grade functions
Improved grade function by adding exception handling

In this version we use overloading to create a second grade function.
The original grade function assumed we had the median score of the
homework grades, rather than the vector containing the individual scores.
The new grade function takes the homework vector as an argument as a
"constant double ref". We also added some basic error handling for the
case where the homework vector is empty.