Reading hw grades
Added new function read_hw() for reading homework scores.

This function accepts input from the user. This function overcomes the
challenge of returning a homework score and an indication of whether the
attempted user input was successful. To do so the function is a parameter
that is a reference to an object in which to place it.