Store student data in struct
Added custom data struct called Student_Info to store student records.

This version is a significant update relative to the previous versions.
The program no longer prompts the user to input the record fields for
each student. The program accepts data in the form "Name midterm final
homework1 homework2 homework3" until an the end-of-file command is recei-
ved . The program then takes the vector containing each student record
and computes a final grade and prints an alphabetized list of students
and their final grade.